refactoring

+++ src/main/java/com/m1namoto/action/DelUserAction.java
@@ -1,16 +1,28 @@
 package com.m1namoto.action;
 
+import com.google.common.base.Optional;
 import com.m1namoto.domain.User;
 import com.m1namoto.page.PageData;
 import com.m1namoto.service.UserService;
 
 public class DelUserAction extends Action {
 
+    private static final String USER_ID_WAS_NOT_PASSED = "User ID was not passed";
+    private static final String INVALID_USER_ID = "Can not parse user id";
+
     @Override
     protected ActionResult execute() throws Exception {
-        String userId = getRequestParamValue("id");
+        Optional<String> userIdOpt = getRequestParamValue("id");
+        if (!userIdOpt.isPresent()) {
+            throw new RuntimeException(USER_ID_WAS_NOT_PASSED);
+        }
+
         User user = new User();
-        user.setId(Long.valueOf(userId));
+        try {
+            user.setId(Long.valueOf(userIdOpt.get()));
+        } catch (Exception e) {
+            throw new RuntimeException(INVALID_USER_ID + ": " + userIdOpt.get());
+        }
         UserService.remove(user);
 
         return createAjaxResult(new PageData());
